pyproject.toml
[project]
name = "pyespn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Steven Morse", email = "steven.t.morse@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
